Make this example more meaningful/fix bug
By doing:

```
@contextlib.contextmanager
def unlock(resource):
    resource = 'unlocked'
    ...
```
We don't really change `resource` that is passed in. Python is pass by object reference so doing a re-assignment does not actually change the `resource`. Furthermore, the `resource` variable, with which the `unlock` function is called, is a string; strings are immutable; thus, we need to pass in a mutable data type as the `resource`. While we could have used a simple "box" tactic i.e. `resource = ['locked']`, this seemed to distract from the goal of the example. Consequently, it seemed using a small class might be clearer without detracting from the example's value.